requirements.txt
geopandas
matplotlib
jinja2
pandas
pdfkit
flask
requests
contextily
selenium
Pillow